Using this as a log, remember to review later
Because the scouting form starts off with all 0s and nulls, setting the teamNumber to 0 might allow the CSV program to immediately delete it. matchNumber should be 0 which could also be used, but I can gaurentee team numbr is not auto-set anywhere. Gonna have to rely on later-filtering unless I can get the server running properly.